# getViewportUniforms Optimization