--- a/src/lib/viewport-uniforms.js
+++ b/src/lib/viewport-uniforms.js
@@ -26,7 +26,7 @@ import {COORDINATE_SYSTEM} from './constants';
 
 function fp64ify(a) {
   const hiPart = Math.fround(a);
-  const loPart = a - Math.fround(a);
+  const loPart = a - hiPart;
   return [hiPart, loPart];
 }