--- a/src/lib/viewport-uniforms.js
+++ b/src/lib/viewport-uniforms.js
@@ -20,6 +20,7 @@
 
 /* global window */
 import {mat4, vec4} from 'gl-matrix';
+// import {Matrix4} from 'luma.gl';
 
 import assert from 'assert';
 import {COORDINATE_SYSTEM} from './constants';
@@ -35,6 +36,9 @@ const ZERO_VECTOR = [0, 0, 0, 0];
 // 4x4 matrix that drops 4th component of vector
 const VECTOR_TO_POINT_MATRIX = [1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0];
 
+// The code that utilizes Matrix4 does the same calculation as their mat4 counterparts,
+// has lower performance but provides error checking.
+// Uncomment when debugging
 function calculateMatrixAndOffset({
   projectionMode,
   positionOrigin,
@@ -51,6 +55,7 @@ function calculateMatrixAndOffset({
   case COORDINATE_SYSTEM.IDENTITY:
   case COORDINATE_SYSTEM.LNGLAT:
     projectionCenter = ZERO_VECTOR;
+    // modelViewMatrix = new Matrix4(viewMatrix);
     modelViewMatrix = mat4.copy([], viewMatrix);
     break;
 
@@ -60,12 +65,16 @@ function calculateMatrixAndOffset({
     // This is the key to offset mode precision (avoids doing this
     // addition in 32 bit precision)
     const positionPixels = viewport.projectFlat(positionOrigin);
+    // projectionCenter = new Matrix4(viewProjectionMatrix)
+    //   .transformVector([positionPixels[0], positionPixels[1], 0.0, 1.0]);
     projectionCenter = vec4.transformMat4([],
       [positionPixels[0], positionPixels[1], 0.0, 1.0],
       viewProjectionMatrix);
 
     // Always apply uncentered projection matrix if available (shader adds center)
     // Zero out 4th coordinate ("after" model matrix) - avoids further translations
+    // modelViewMatrix = new Matrix4(viewMatrixUncentered || viewMatrix)
+    //   .multiplyRight(VECTOR_TO_POINT_MATRIX);
     modelViewMatrix = mat4.multiply([], viewMatrixUncentered || viewMatrix, VECTOR_TO_POINT_MATRIX);
     break;
 
@@ -77,9 +86,11 @@ function calculateMatrixAndOffset({
 
   if (modelMatrix) {
     // Apply model matrix if supplied
+    // modelViewMatrix.multiplyRight(modelMatrix);
     mat4.multiply(modelViewMatrix, modelViewMatrix, modelMatrix);
   }
 
+  // const modelViewProjectionMatrix = new Matrix4(projectionMatrix).multiplyRight(modelViewMatrix);
   const modelViewProjectionMatrix = mat4.multiply([], projectionMatrix, modelViewMatrix);
   const cameraPos = [viewMatrixInv[12], viewMatrixInv[13], viewMatrixInv[14]];