--- a/test/data/index.js
+++ b/test/data/index.js
@@ -19,8 +19,10 @@
 // THE SOFTWARE.
 
 export * from '../../examples/layer-browser/src/data-samples';
+export * from './viewport';
 
 import * as data from '../../examples/layer-browser/src/data-samples';
+export const lines = data.choropleths.features.map(f => ({path: f.geometry.coordinates[0]}));
 
 import Immutable from 'immutable';
 export const immutableChoropleths = Immutable.fromJS(data.choropleths);