Changes:
- Switch from using custom math class to `gl-matrix`
- Add benchmark

Before:
```
getUniformsFromViewport#LNGLAT x 54,286 ops/sec ±2.77% (84 runs sampled)
getUniformsFromViewport#METER_OFFSETS x 37,155 ops/sec ±2.44% (84 runs sampled)
```
After:
```
getUniformsFromViewport#LNGLAT x 155,845 ops/sec ±1.70% (88 runs sampled)
getUniformsFromViewport#METER_OFFSETS x 123,868 ops/sec ±0.86% (90 runs sampled)
```