@Pessimistress 2017-04-21
With `glMatrix.debug` off:
```
getUniformsFromViewport#LNGLAT x 54,471 ops/sec ±5.73% (81 runs sampled)
getUniformsFromViewport#METER_OFFSETS x 40,144 ops/sec ±2.07% (86 runs sampled)
```