@Pessimistress 2017-04-21
> Still not totally happy since I really value the readability and debug support of the math classes.

My take is that the math classes are great when we are constructing reusable props, but for the code that runs every frame we should prioritize performance. The Viewport classes use gl-matrix directly, for example.