@ibgreen 2017-04-21
> With `glMatrix.debug` off:
OK thanks. That was a surprisingly small effect. Will need to port the perf framework to luma.gl and add this as a test case.