@ibgreen 2017-04-21
> My take is that the math classes are great when we are constructing reusable props, but for the code that runs every frame we should prioritize performance. 

Yes, agreed, all is not lost. The Math classes still have a role to play.

But regardless, I would expect the classes to be slower by maybe 20%, but not by a factor of 3x. Perhaps there is something that pulls down performance that can be tweaked.