@shaojingli 2017-04-23
> But regardless, I would expect the classes to be slower by maybe 20%, but not by a factor of 3x. Perhaps there is something that pulls down performance that can be tweaked.

This is seems to be due to SIMD acceleration of those linear algebra operations