@ibgreen 2017-04-24
> This is seems to be due to SIMD acceleration of those linear algebra operations

Yes `gl-matrix` does have SIMD acceleration. That said, the math classes use `gl-matrix` underneath the hood, so if `gl-matrix` SIMD acceleration is active for the basic library, it should be active for the class wrappers as well. 

On this topic, current impression is that Firefox supports the SIMD extensions but Chrome doesn't just yet - or maybe that has changed?

Finally @Pessimistress did some profiling and it looks like it is a transpilation issue - we are using the ES6 feature of inheriting from JavaScript `Array`s (instead of `Object`s) which is not efficiently poly filled in ES5 environments.