@ibgreen 2017-04-24
As stated before, I would just like to see the Math class code preserved in the comments rather than deleted completely so that I can activate it when debugging the viewport, rather than having to reimplement it again.

As usual when keeping commented code, there needs to be a comment explaining why we keep it.